# src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Austin Rifle Club API",
    "description": "API for Austin Rifle Club member portal, range status, events, and administration.",
    "version": "1.0.0",
    "contact": {
      "email": "info@austinrifleclub.org"
    }
  },
  "servers": [
    {
      "url": "https://austinrifleclub.org/api",
      "description": "Production"
    },
    {
      "url": "http://localhost:8787/api",
      "description": "Development"
    }
  ],
...
  "paths": {
...
    "/events": {
      "get": {
        "tags": ["Events"],
        "summary": "List events",
        "description": "Returns published events. Non-authenticated users see only public events.",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "start", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "end", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "type", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Event list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "events": { "type": "array", "items": { "$ref": "#/components/schemas/Event" } },
                    "pagination": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Events"],
        "summary": "Create event (admin only)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateEvent" }
            }
          }
        },
        "responses": {
          "201": { "description": "Event created" }
        }
      }
    },
    "/events/{id}": {
      "get": {
        "tags": ["Events"],
        "summary": "Get event details",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Event details" },
          "404": { "description": "Event not found" }
        }
      },
      "patch": {
        "tags": ["Events"],
        "summary": "Update event (admin only)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Event updated" }
        }
      }
    },
    "/events/{id}/register": {
      "post": {
        "tags": ["Events"],
        "summary": "Register for event (member only)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "division": { "type": "string" },
                  "classification": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Registration confirmed or waitlisted" },
          "400": { "description": "Registration not open or deadline passed" },
          "409": { "description": "Already registered" }
        }
      },
      "delete": {
        "tags": ["Events"],
        "summary": "Cancel event registration (member only)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Registration cancelled with refund info" },
          "404": { "description": "Not registered" }
        }
      }
    },
...
    "/payments/checkout/event/{eventId}": {
      "post": {
        "tags": ["Payments"],
        "summary": "Create event registration checkout session",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "eventId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Checkout session created" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "Event": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "eventType": { "type": "string" },
          "startTime": { "type": "string", "format": "date-time" },
          "endTime": { "type": "string", "format": "date-time" },
          "location": { "type": "string" },
          "maxParticipants": { "type": "integer", "nullable": true },
          "cost": { "type": "number", "nullable": true },
          "isPublic": { "type": "boolean" },
          "registrationCount": { "type": "integer" },
          "spotsRemaining": { "type": "integer", "nullable": true }
        }
      },
      "CreateEvent": {
        "type": "object",
        "required": ["title", "startTime", "endTime"],
        "properties": {
          "title": { "type": "string" },
          "description": { "type": "string" },
          "eventType": { "type": "string" },
          "startTime": { "type": "string", "format": "date-time" },
          "endTime": { "type": "string", "format": "date-time" },
          "location": { "type": "string" },
          "maxParticipants": { "type": "integer" },
          "cost": { "type": "number" },
          "isPublic": { "type": "boolean" },
          "rangeIds": { "type": "array", "items": { "type": "string" } }
        }
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "total": { "type": "integer" },
          "totalPages": { "type": "integer" }
        }
      }
    }
  }
}